added cart to other pages
